[gui] Add indicators for field constraints and default values in the attributes form designer (partial)
