+++ src/gui/vector/qgsattributesformtreeviewindicator.cpp
@@ -1,0 +1,6 @@
+#include "qgsattributesformtreeviewindicator.h"
+
+QgsAttributesFormTreeViewIndicator::QgsAttributesFormTreeViewIndicator( QObject *parent )
+  : QObject { parent }
+{
+}
